biaslens-ai/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
fairlearn>=0.11.0
aif360>=0.6.1
joblib>=1.4.0
pydantic>=1.10.0,<2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
